Improve avatar fetching and code organization
